Use ingredient font for all ingredient tooltips
