+++ src/main/java/mezz/jei/gui/ingredients/GuiIngredient.java
@@ -1,6 +1,35 @@
 package mezz.jei.gui.ingredients;
 
 import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.systems.RenderSystem;
+import mezz.jei.Internal;
+import mezz.jei.api.gui.drawable.IDrawable;
+import mezz.jei.api.gui.ingredient.IGuiIngredient;
+import mezz.jei.api.gui.ingredient.ITooltipCallback;
+import mezz.jei.api.helpers.IModIdHelper;
+import mezz.jei.api.ingredients.IIngredientHelper;
+import mezz.jei.api.ingredients.IIngredientRenderer;
+import mezz.jei.api.ingredients.subtypes.UidContext;
+import mezz.jei.api.recipe.IFocus;
+import mezz.jei.gui.Focus;
+import mezz.jei.gui.TooltipRenderer;
+import mezz.jei.ingredients.IngredientFilter;
+import mezz.jei.render.IngredientRenderHelper;
+import mezz.jei.util.ErrorUtil;
+import net.minecraft.client.gui.AbstractGui;
+import net.minecraft.client.renderer.Rectangle2d;
+import net.minecraft.client.renderer.RenderHelper;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.tags.ITag;
+import net.minecraft.tags.ITagCollection;
+import net.minecraft.tags.ItemTags;
+import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraft.util.text.TranslationTextComponent;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
@@ -9,39 +38,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-
-import com.mojang.blaze3d.systems.RenderSystem;
-import mezz.jei.api.ingredients.subtypes.UidContext;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.AbstractGui;
-import net.minecraft.client.gui.FontRenderer;
-import net.minecraft.client.renderer.Rectangle2d;
-import net.minecraft.client.renderer.RenderHelper;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.tags.ItemTags;
-import net.minecraft.tags.ITagCollection;
-import net.minecraft.tags.ITag;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextFormatting;
-
-import mezz.jei.Internal;
-import mezz.jei.api.gui.drawable.IDrawable;
-import mezz.jei.api.gui.ingredient.IGuiIngredient;
-import mezz.jei.api.gui.ingredient.ITooltipCallback;
-import mezz.jei.api.helpers.IModIdHelper;
-import mezz.jei.api.ingredients.IIngredientHelper;
-import mezz.jei.api.ingredients.IIngredientRenderer;
-import mezz.jei.api.recipe.IFocus;
-import mezz.jei.gui.Focus;
-import mezz.jei.gui.TooltipRenderer;
-import mezz.jei.ingredients.IngredientFilter;
-import mezz.jei.render.IngredientRenderHelper;
-import mezz.jei.util.ErrorUtil;
-import net.minecraft.util.text.TranslationTextComponent;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 public class GuiIngredient<T> extends AbstractGui implements IGuiIngredient<T> {
 	private static final Logger LOGGER = LogManager.getLogger();
@@ -222,8 +218,6 @@ public class GuiIngredient<T> extends AbstractGui implements IGuiIngredient<T> {
 				tooltipCallback.onTooltip(slotIndex, input, value, tooltip);
 			}
 
-			Minecraft minecraft = Minecraft.getInstance();
-			FontRenderer fontRenderer = ingredientRenderer.getFontRenderer(minecraft, value);
 			if (value instanceof ItemStack) {
 				//noinspection unchecked
 				Collection<ItemStack> itemStacks = (Collection<ItemStack>) this.allIngredients;
@@ -233,7 +227,7 @@ public class GuiIngredient<T> extends AbstractGui implements IGuiIngredient<T> {
 					tooltip.add(acceptsAny.withStyle(TextFormatting.GRAY));
 				}
 			}
-			TooltipRenderer.drawHoveringText(value, tooltip, xOffset + mouseX, yOffset + mouseY, fontRenderer, matrixStack);
+			TooltipRenderer.drawHoveringText(value, tooltip, xOffset + mouseX, yOffset + mouseY, matrixStack);
 
 			RenderSystem.enableDepthTest();
 		} catch (RuntimeException e) {
